+++ internal/hotkey/modifiers_expand_default.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package hotkey
+
+import "golang.design/x/hotkey"
+
+func expandModifiers(modifiers []hotkey.Modifier) [][]hotkey.Modifier {
+	return [][]hotkey.Modifier{modifiers}
+}